admin/pyproject.toml
[tool.poetry]
name = "bot admin"
version = "0.1.0"
description = "admin for photo bot"
authors = ["Valeriya Sidorova <va_sidorova@inbox.ru>", "optikrus <optikrus61@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
Django = "^4.2.5"
django-environ = "^0.11.2"
psycopg2-binary = "2.9.9"
python-dotenv = "1.0.0"
ruff = "^0.1.6"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
